Moved outdated functions to pyflextrkr/depreciated directory.
1. Updated UserGuide.md to reflect the new changes.

2. Moved previously separate map feature functions to the pyflextrkr/depreciated directory.
