+++ pyflextrkr/depreciated/mapcell_radar.py
