+++ pyflextrkr/depreciated/mapgeneric.py
