+++ runscripts/run_mcsdata_allradarmatch.py
@@ -454,7 +454,7 @@ if run_labelmcs == 1:
     # Process files
 
     # Load function 
-    from pyflextrkr.mapmcs import mapmcs_pf
+    from pyflextrkr.depreciated.mapmcs import mapmcs_pf
 
     # Generate input list
     list_robustmcsstat_filebase = [robustmcs_filebase]*(cloudidfilestep-1)
